# Conflicts:
#	ext/pdo_sqlite/config.m4
#	ext/pdo_sqlite/config.w32

+++ ext/hash/hash_joaat.c
@@ -16,8 +16,6 @@
   +----------------------------------------------------------------------+
 */
 
-/* $Id$ */
-
 /* Implements Jenkins's one-at-a-time hashing algorithm as presented on
  * http://www.burtleburtle.net/bob/hash/doobs.html.
  */
@@ -41,7 +39,7 @@ PHP_HASH_API void PHP_JOAATInit(PHP_JOAAT_CTX *context)
 	context->state = 0;
 }
 
-PHP_HASH_API void PHP_JOAATUpdate(PHP_JOAAT_CTX *context, const unsigned char *input, unsigned int inputLen)
+PHP_HASH_API void PHP_JOAATUpdate(PHP_JOAAT_CTX *context, const unsigned char *input, size_t inputLen)
 {
 	context->state = joaat_buf((void *)input, inputLen, context->state);
 }
