+++ ext/hash/php_hash_ripemd.h
@@ -16,8 +16,6 @@
    +----------------------------------------------------------------------+
 */
 
-/* $Id$ */
-
 #ifndef PHP_HASH_RIPEMD_H
 #define PHP_HASH_RIPEMD_H
 #include "ext/standard/basic_functions.h"
@@ -48,19 +46,19 @@ typedef struct {
 } PHP_RIPEMD320_CTX;
 
 PHP_HASH_API void PHP_RIPEMD128Init(PHP_RIPEMD128_CTX *);
-PHP_HASH_API void PHP_RIPEMD128Update(PHP_RIPEMD128_CTX *, const unsigned char *, unsigned int);
+PHP_HASH_API void PHP_RIPEMD128Update(PHP_RIPEMD128_CTX *, const unsigned char *, size_t);
 PHP_HASH_API void PHP_RIPEMD128Final(unsigned char[16], PHP_RIPEMD128_CTX *);
 
 PHP_HASH_API void PHP_RIPEMD160Init(PHP_RIPEMD160_CTX *);
-PHP_HASH_API void PHP_RIPEMD160Update(PHP_RIPEMD160_CTX *, const unsigned char *, unsigned int);
+PHP_HASH_API void PHP_RIPEMD160Update(PHP_RIPEMD160_CTX *, const unsigned char *, size_t);
 PHP_HASH_API void PHP_RIPEMD160Final(unsigned char[20], PHP_RIPEMD160_CTX *);
 
 PHP_HASH_API void PHP_RIPEMD256Init(PHP_RIPEMD256_CTX *);
-PHP_HASH_API void PHP_RIPEMD256Update(PHP_RIPEMD256_CTX *, const unsigned char *, unsigned int);
+PHP_HASH_API void PHP_RIPEMD256Update(PHP_RIPEMD256_CTX *, const unsigned char *, size_t);
 PHP_HASH_API void PHP_RIPEMD256Final(unsigned char[32], PHP_RIPEMD256_CTX *);
 
 PHP_HASH_API void PHP_RIPEMD320Init(PHP_RIPEMD320_CTX *);
-PHP_HASH_API void PHP_RIPEMD320Update(PHP_RIPEMD320_CTX *, const unsigned char *, unsigned int);
+PHP_HASH_API void PHP_RIPEMD320Update(PHP_RIPEMD320_CTX *, const unsigned char *, size_t);
 PHP_HASH_API void PHP_RIPEMD320Final(unsigned char[40], PHP_RIPEMD320_CTX *);
 
 #endif /* PHP_HASH_RIPEMD_H */
